requirements.txt
discord.py
dislash.py
translate
pipe
tokenizer
googletrans
python-dotenv
mecab-python3
pytz
spacy
chatterbot-corpus 
ChatterBot